requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-ai==0.0.13
python-dotenv==1.0.0
httpx==0.26.0